requirements.txt
# Core uAgents framework and dependencies
uagents>=0.22.9
uagents-core>=0.3.9

# HTTP client for API calls
httpx>=0.28.1

# Data validation and serialization
pydantic>=2.11.9

# Environment variable management
python-dotenv>=1.1.1

# Blockchain and cryptography
cosmpy>=0.11.1
ecdsa>=0.19.1
bech32>=1.2.0

# Async support
aiohttp>=3.12.15
uvicorn>=0.37.0

# For demo only - not required for agents
asyncio-mqtt>=0.16.1